backend/requirements.txt
fastapi
uvicorn
edge-tts
httpx
python-multipart
aiofiles
llama-cpp-python
